Day 17 Part 2 stolen solution
I for the life of me could not get the correct number of valid entries
in part 2.
I have not yet grasped all the tricks that can be used, nor have I
mastered Python, either.
But I have been able to get this far without directly using
someone else's code.
So, that is a plus?